apps/backend/pyproject.toml
[tool.poetry]
name = "codexos-backend"
version = "1.0.0"
description = "CodexOS Backend API - Autonomous Engineering OS"
authors = ["CodexOS Team <team@codexos.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.26.0"
openai = "^1.10.0"
anthropic = "^0.16.0"
langchain = "^0.1.4"
langchain-openai = "^0.0.5"
chromadb = "^0.4.22"
pypdf = "^3.17.4"
beautifulsoup4 = "^4.12.3"
aiofiles = "^23.2.1"
websockets = "^12.0"
celery = {extras = ["redis"], version = "^5.3.4"}
stripe = "^7.12.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.0"}
cryptography = "^42.0.0"
boto3 = "^1.34.25"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
prometheus-client = "^0.19.0"
locust = "^2.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C90", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
